move validator-restapi into validator client module to avoid cyclic deps.
