+++ settings.gradle
@@ -57,7 +57,6 @@ include 'validator:client'
 include 'validator:eventadapter'
 include 'validator:relaypublisher'
 include 'validator:remote'
-include 'validator:restapi'
 include 'eth-tests'
 include 'eth-reference-tests'
 include 'eth-benchmark-tests'
