+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/restapi/apis/GetKeys.java
@@ -11,26 +11,29 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.validator.restapi.apis;
+package tech.pegasys.teku.validator.client.restapi.apis;
 
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_INTERNAL_SERVER_ERROR;
+import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_OK;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.EndpointMetadata;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiEndpoint;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiRequest;
+import tech.pegasys.teku.validator.client.restapi.ValidatorTypes;
 
-public class PostKeys extends RestApiEndpoint {
+public class GetKeys extends RestApiEndpoint {
 
   public static final String ROUTE = "/eth/v1/keystores";
 
-  public PostKeys() {
+  public GetKeys() {
     super(
-        EndpointMetadata.post(ROUTE)
-            .operationId("ImportKeystores")
-            .summary("Import Keystores")
-            .description("Import keystores generated by the Eth2.0 deposit CLI tooling.")
-            // TODO add request and response
+        EndpointMetadata.get(ROUTE)
+            .operationId("ListKeys")
+            .summary("List Keys")
+            .description(
+                "List all validating pubkeys known to and decrypted by this keymanager binary")
+            .response(SC_OK, "Success response", ValidatorTypes.LIST_KEYS_RESPONSE_TYPE)
             .withAuthenticationResponses()
             .build());
   }
