+++ validator/client/src/test/resources/tech/pegasys/teku/validator/client/restapi/PubKey.json
