--- validator/restapi/build.gradle
@@ -1,28 +1,0 @@
-dependencies {
-  implementation project(':bls')
-  implementation project(':infrastructure:async')
-  implementation project(':infrastructure:events')
-  implementation project(':infrastructure:exceptions')
-  implementation project(':infrastructure:http')
-  implementation project(':infrastructure:restapi')
-  implementation project(':infrastructure:version')
-  implementation project(':ethereum:core')
-  implementation project(':ethereum:spec')
-  implementation project(':data:serializer')
-  implementation project(':util')
-
-  implementation 'com.fasterxml.jackson.module:jackson-module-kotlin'
-  implementation 'com.google.guava:guava'
-  implementation 'io.swagger.core.v3:swagger-core'
-  implementation 'io.swagger.core.v3:swagger-annotations'
-  implementation 'io.github.classgraph:classgraph'
-  implementation 'io.javalin:javalin'
-  implementation 'io.javalin:javalin-openapi'
-  implementation 'org.apache.tuweni:tuweni-bytes'
-  implementation 'org.webjars:swagger-ui'
-
-  testImplementation testFixtures(project(':bls'))
-  testImplementation testFixtures(project(':ethereum:spec'))
-  testImplementation testFixtures(project(':infrastructure:restapi'))
-}
-
